requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
datasets==3.1.0
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-aiplatform==1.73.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.12.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.3
idna==3.10
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.9
langchain-anthropic==0.3.0
langchain-core==0.3.33
langchain-deepseek==0.1.1
langchain-google-vertexai==2.0.8
langchain-groq==0.2.1
langchain-mistralai==0.2.3
langchain-openai==0.3.3
langchain-text-splitters==0.3.2
langsmith==0.1.147
multidict==6.1.0
multiprocess==0.70.16
numpy==1.26.4
openai==1.61.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.6.0
scipy==1.15.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.18.3
